test: 동시성 테스트에 예외 테스트 추가
